modify: folder structure + navbar
