+++ src/App.js
@@ -3,8 +3,8 @@ import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import Mainpage from './components/homepage/Mainpage';
 import SignIn from './components/auth/SignIn';
 import SignUp from './components/auth/SignUp';
-import Navbar from './components/layout/Navbar';
-import AddPlayers from './select players/AddPlayers';
+import Navbar from './components/navbar/NavbarContainer';
+import AddPlayers from './components/select players/AddPlayers';
 import Contests from "./components/contests";
 import pools from "./components/pools";
 
@@ -14,10 +14,9 @@ class App extends React.Component {
   return (
     <BrowserRouter>
     <div>
-      <Navbar />
+      <Navbar/>
         <Switch>
           <Route exact path='/' component={Mainpage} />
-          {/* <Route path='/signin' component={SignIn} /> */}
           <Route path='/signup' component={SignUp }/>
           <Route path='/batsmen' component={AddPlayers} />
           <Route path="/CricketPool1" component={pools} />
