--- src/components/layout/Navbar.js
@@ -1,20 +1,0 @@
-import React from 'react';
-import {Link} from 'react-router-dom';
-import SignedInLinks from './SignedInLinks';
-import SignedOutLinks from './SIgnedOutLinks';
-
-const Navbar=()=> {
-    return(
-        <nav className="nav-wrapper grey darken-3">
-            <div className="container">
-                <Link to='/' className="brand-logo">
-                    XPL
-                </Link>
-                <SignedInLinks/>
-                <SignedOutLinks/>
-            </div> 
-        </nav>
-    )
-}
-
-export default Navbar;
