+++ src/components/navbar/SIgnedOutLinks.js
@@ -1,13 +1,12 @@
 import React from 'react';
-import {NavLink} from 'react-router-dom';
 import './SignedOutLinks.css';
-import ModalLauncher from '../../landing-page/ModalLauncher';
+import ModalLauncher from '../sign_in/SignInModalLauncher';
 
 const SignedOutLinks=()=> {
     return(
         <ul className="right">
             {/* <li><NavLink to="/signup" exact>Sign Up</NavLink> </li> */}
-            <li><button><ModalLauncher/></button> </li>
+            <li><ModalLauncher/></li>
         </ul>
     )
 }
