+++ src/components/select players/PlayerCard.css
