+++ src/components/select players/PlayerCard.js
