+++ src/components/select players/SelectPlayerNavbar.css
