+++ src/components/sign_in/auth_modal/Heading.js
