+++ src/store/reducers/rootreducer.js
@@ -1,10 +1,10 @@
 import authReducer from './authReducer';
-import modalReducer from './teamReducer';
+import teamReducer from './teamReducer';
 import {combineReducers} from 'redux';
 
 const rootReducer = combineReducers({
     auth: authReducer,
-    modal: modalReducer
+    team: teamReducer
 });
 
 export default rootReducer;
